+++ cmd/empire/factories.go
@@ -74,6 +74,13 @@ func newEmpire(db *empire.DB, c *Context) (*empire.Empire, error) {
 	e.Environment = c.String(FlagEnvironment)
 	e.RunRecorder = runRecorder
 	e.MessagesRequired = c.Bool(FlagMessagesRequired)
+
+	switch c.String(FlagAllowedCommands) {
+	case "procfile":
+		e.AllowedCommands = empire.AllowCommandProcfile
+	default:
+	}
+
 	if logs != nil {
 		e.LogsStreamer = logs
 	}