+++ cmd/empire/main.go
@@ -10,13 +10,15 @@ import (
 )
 
 const (
-	FlagPort             = "port"
-	FlagAutoMigrate      = "automigrate"
-	FlagScheduler        = "scheduler"
-	FlagEventsBackend    = "events.backend"
-	FlagRunLogsBackend   = "runlogs.backend"
+	FlagPort           = "port"
+	FlagAutoMigrate    = "automigrate"
+	FlagScheduler      = "scheduler"
+	FlagEventsBackend  = "events.backend"
+	FlagRunLogsBackend = "runlogs.backend"
+	FlagLogLevel       = "log.level"
+
 	FlagMessagesRequired = "messages.required"
-	FlagLogLevel         = "log.level"
+	FlagAllowedCommands  = "commands.allowed"
 
 	FlagStats = "stats"
 
@@ -346,6 +348,12 @@ var EmpireFlags = []cli.Flag{
 		Name:   FlagMessagesRequired,
 		Usage:  "If true, messages will be required for empire actions that emit events.",
 		EnvVar: "EMPIRE_MESSAGES_REQUIRED",
+	},
+	cli.StringFlag{
+		Name:   FlagAllowedCommands,
+		Value:  "any",
+		Usage:  "Specifies what commands are allowed when using `emp run`. Can be `any`, or `procfile`.",
+		EnvVar: "EMPIRE_ALLOWED_COMMANDS",
 	},
 	cli.BoolFlag{
 		Name:   FlagXShowAttached,