+++ extractor.go
@@ -242,8 +242,9 @@ func formationFromExtendedProcfile(p procfile.ExtendedProcfile) (Formation, erro
 		}
 
 		f[name] = Process{
-			Command: cmd,
-			Cron:    process.Cron,
+			Command:   cmd,
+			Cron:      process.Cron,
+			NoService: process.NoService,
 		}
 	}