+++ processes.go
@@ -4,6 +4,7 @@ import (
 	"database/sql/driver"
 	"encoding/json"
 	"errors"
+	"fmt"
 	"strings"
 
 	shellwords "github.com/mattn/go-shellwords"
@@ -70,6 +71,10 @@ type Process struct {
 	// Command is the command to run.
 	Command Command `json:"Command,omitempty"`
 
+	// Signifies that this is a named one off command and not a long lived
+	// service.
+	NoService bool `json:"Run,omitempty"`
+
 	// Quantity is the desired number of instances of this process.
 	Quantity int `json:"Quantity,omitempty"`
 
@@ -85,6 +90,18 @@ type Process struct {
 	// A cron expression. If provided, the process will be run as a
 	// scheduled task.
 	Cron *string `json:"cron,omitempty"`
+}
+
+// IsValid returns nil if the Process is valid.
+func (p *Process) IsValid() error {
+	// Ensure that processes marked as NoService can't be scaled up.
+	if p.NoService {
+		if p.Quantity != 0 {
+			return errors.New("non-service processes cannot be scaled up")
+		}
+	}
+
+	return nil
 }
 
 // Constraints returns a constraints.Constraints from this Process definition.
@@ -106,6 +123,17 @@ func (p *Process) SetConstraints(c Constraints) {
 
 // Formation represents a collection of named processes and their configuration.
 type Formation map[string]Process
+
+// IsValid returns nil if all of the Processes are valid.
+func (f Formation) IsValid() error {
+	for n, p := range f {
+		if err := p.IsValid(); err != nil {
+			return fmt.Errorf("process %s is not valid: %v", n, err)
+		}
+	}
+
+	return nil
+}
 
 // Scan implements the sql.Scanner interface.
 func (f *Formation) Scan(src interface{}) error {